requirements.txt
pytest==7.4.0
python-dotenv==1.0.0
fastapi==0.68.0
uvicorn==0.15.0
redis==4.3.4
pyyaml==6.0 